Restructure docs components folder
