+++ docs/src/components/docs/index.ts
@@ -1,0 +1,6 @@
+export { default as DocsComponents } from './DocsComponents';
+export { default as DocsHead } from './DocsHead';
+export { default as DocsPage } from './DocsPage';
+export { default as DocsPlaygroundControls } from './DocsPlaygroundControls';
+export { default as DocsPlaygroundControl } from './DocsPlaygroundControl';
+export { default as DocsPlayground } from './DocsPlayground';
