+++ docs/src/components/icons.ts
@@ -1,0 +1,2 @@
+export { default as Add } from 'assets/icons/add.svg';
+export { default as ChevronRight } from 'assets/icons/chevron_right.svg';
