+++ docs/src/pages/components/button.mdx
@@ -6,9 +6,8 @@ description: Buttons allow users to perform actions and make choices when presse
 import { Fragment, useState } from 'react';
 import { Button } from '@theomer77/omer-ui-components';
 
-import DocsPlayground from 'components/DocsPlayground';
-import DocsComponents from 'components/DocsComponents';
-import AddIcon from 'assets/icons/add.svg';
+import { DocsComponents, DocsPlayground } from 'components/docs';
+import { Add as AddIcon } from 'components/icons';
 
 export const buttonSizes = [
     { label: 'Extra small', value: 'xs' },
